% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{add_col}
\alias{add_col}
\title{Safe column addition}
\usage{
add_col(.data, ...)
}
\description{
\code{add_col()} adds a column(s) into the data frame \code{x}
if the column(s) is/are not present. The way it is written is
intended to work both for standard \code{data.frame} and other
\code{data.frame} derived classes (such as \code{dplyr::tibble()}) provided
the derived class has the method \verb{[<-} defined.
}
\examples{
# works for data.frames and tibbles
x = data.frame(id=1:2)
add_col(x, name=c('John','Jeff'), age=c(20,40))
#add_col(as_tibble(x), name=c('John','Jeff'), age=c(20,40))

}
\seealso{
Other matrix/data-frame functions provided by ub: 
\code{\link{anti_join_sym}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{rev_rows}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by ub}
