% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_tight_axes}
\alias{add_tight_axes}
\title{Add tight axis labels to a plot}
\usage{
add_tight_axes(
  xlab = NULL,
  ylab = NULL,
  linex = 1.4,
  liney = 1.47,
  padjx = -0.85,
  padjy = 0.85,
  tclx = -0.3,
  tcly = -0.3,
  cex = NULL,
  cexx = 1,
  cexy = 1,
  cexfaclabx = 1,
  cexfaclaby = 1,
  xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf),
  scipenx,
  scipeny,
  scipen0,
  ylabfun,
  xlabfun,
  do_x = TRUE,
  do_y = TRUE
)
}
\description{
Add tight axis labels to a plot
}
\seealso{
Other plotting-related functions provided by ub: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by ub}
