% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{all_devs}
\alias{all_devs}
\alias{.all_devs}
\alias{.all_devs_off}
\alias{.ado}
\title{List/close graphical devices}
\usage{
.all_devs(silent = TRUE)

.all_devs_off(silent = TRUE)

.ado()
}
\description{
List or close all graphical devices.

\code{.all_devs} returns a list of all open graphical devices with an
announcement of the number of open devices. The announcement can be
disabled.

\code{.all_devs_off} and \code{.ado} close all open graphical devices either
with announcement (\code{.ado}) or silently (\code{.all_devs_off}).
}
\examples{
.all_devs(silent=FALSE)

}
\seealso{
Other plotting-related functions provided by ub: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}

Other plotting-related functions provided by ub: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}

Other plotting-related functions provided by ub: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by ub}
