% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{anyNonNA}
\alias{anyNonNA}
\title{Check for non-NA values}
\usage{
anyNonNA(x, rowwise = TRUE)
}
\description{
\code{anyNonNA()} checks whether a row of \code{x} contains at least one
non-NA value. Works for vectors and 2-arrays.
}
\examples{
x = rbind(c(1,NA),c(NA,NA)); anyNonNA(x)

}
\seealso{
Other check-performing functions provided by ub: 
\code{\link{extract_by_positiveness_negativeness}},
\code{\link{is_number}()},
\code{\link{is_same_value}()},
\code{\link{numchecks}},
\code{\link{rounding}}
}
\concept{check-performing functions provided by ub}
