% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{ask_enter_with_countdown}
\alias{ask_enter_with_countdown}
\title{Read input with a countdown}
\usage{
ask_enter_with_countdown(
  n = 10,
  msg = "Press <ENTER> or continuing in {i} seconds ..."
)
}
\description{
\code{ask_enter_with_countdown()} asks for ENTER key for \code{n} seconds while
showing a coundown from \code{n} down to \code{1} after which point it exits
returning either the user input as character or \code{NULL} (on no input).
}
\examples{
x = ask_enter_with_countdown(3)

}
