% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorchecks.R
\name{attemp_call}
\alias{attemp_call}
\title{Attempt to call}
\usage{
attemp_call(expr)
}
\description{
Make an attempt to call a supplied call and evaluate whether an
error occurred.
}
\examples{
attemp_call(sin(0))     # call runs ok, returns TRUE
attemp_call(sin('a'))   # call fails, returns FALSE

}
