% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_package}
\alias{build_package}
\alias{build_ub}
\alias{install_ub}
\title{Build a package file}
\usage{
build_package(
  pckg_name,
  pckg_dir = pckg_name,
  pckg_dir_source = "source",
  path = ".",
  create = TRUE,
  update_doc = TRUE,
  build = TRUE,
  install_mode = c("quick", "full", "no"),
  file_desc = "DESCRIPTION",
  increase_version = c("minor3", "minor2", "minor1", "major", "none"),
  add_path_to_global = TRUE,
  attach = TRUE,
  install_prerequisities = TRUE
)

build_ub(
  pckg_name = "ub",
  pckg_dir = "ub",
  pckg_dir_source = "source",
  path = ".",
  file_desc = "DESCRIPTION",
  create = TRUE,
  update_doc = TRUE,
  build = TRUE,
  install_mode = "full",
  attach = FALSE,
  add_path_to_global = TRUE,
  increase_version = "minor3",
  install_prerequisities = TRUE
)

install_ub(
  pckg_name = "ub",
  pckg_dir = "ub",
  path = ".",
  install_mode = "full",
  attach = FALSE,
  add_path_to_global = FALSE,
  install_prerequisities = TRUE,
  pckg_dir_source = "source"
)
}
\description{
\code{build_package()} takes a source code of a package and created the installation file
for the package using the devtools package. Depending on the way it is called, the
the package version in the DESCRIPTION file is automatically increased (controlled by
\code{increase_version}), the documentation is updated (when \code{update_doc = TRUE}) using the
roxygen2 package (via \code{devtools::document()}), the installation file is build (when
\code{build = TRUE}) and the package is installed (regularly when \code{install_mode = 'full'},
or quickly when \code{install_mode = 'quick'}). Optionally, the package installation path
is added to .GlobalEnv and to the .Rprofile file.
}
\details{
\code{build_ub()} calls \code{build_package()} using the settings for 'ub' for a complete
creation (i.e., create, update documentation, build, install, attach)

\code{install_ub()} calls \code{build_package()} using the settings for 'ub'.
}
