% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{call_info_old}
\alias{call_info_old}
\title{Get the details about where a call came from}
\usage{
call_info_old(print_msg = TRUE, level = 1)
}
\value{
A list with the following elements:
\code{fun} ... calling function
\code{dir} ... path to the script file
\code{file} ... name of the script file
\code{line} ... line in the script where called
\code{msg} ... message with the sequence
}
\description{
\code{caller_info_old()} returns a list with information about the calling sequence.
Optionally, a message with the information is directly printed.
}
\details{
\code{call_info_old()} prints and returns the information about the calling sequence
(obtained using \code{caller_info()}.
}
