% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{caller_info}
\alias{caller_info}
\title{Get the details about where a call came from}
\usage{
caller_info(print_msg = FALSE, level = 1, warn = TRUE)
}
\value{
A list with the following elements:
\code{fun} ... calling function
\code{dir} ... path to the script file
\code{file} ... name of the script file
\code{line} ... line in the script where called
\code{message} ... message with the sequence
}
\description{
\code{caller_info()} returns a list with information about the calling sequence.
Optionally, a message with the information is directly printed.
}
\concept{halting utilities provided by ub}
