% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_abbrevstr}
\alias{class_abbrevstr}
\alias{print.abbrevstr}
\title{Methods for class \"\code{abbrevstr}\"}
\usage{
\method{print}{abbrevstr}(x)
}
\description{
Printing method for class \"\code{abbrevstr}\", which extends
\code{\link[base:print]{base::print}}.
}
