% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_huge}
\alias{class_huge}
\alias{is.huge}
\alias{as.huge}
\alias{as.huge.list}
\alias{as.huge.default}
\alias{[.huge}
\alias{[[.huge}
\alias{print.huge}
\alias{print.huge.character}
\alias{print.huge.default}
\alias{truncate_huge}
\alias{truncate_huge.default}
\title{Convert objects to class \code{huge}}
\usage{
is.huge(x)

as.huge(...)

\method{as.huge}{list}(x)

\method{as.huge}{default}(x)

\method{[}{huge}(x, i)

\method{[[}{huge}(x, i)

print.huge(...)

\method{print}{huge.character}(x, n, len, show_note = TRUE)

\method{print}{huge.default}(x, n, len, show_note = TRUE)

truncate_huge(...)

\method{truncate_huge}{default}(x, n)
}
\description{
Class \code{huge} is intended to be used for objects with many elements
which, if all printed, would clutter the screen. With class \code{huge}
set, only a limited number of elements are actually printed.
(extracted via \code{base::head()}).

\code{is_huge()} checks whether class \code{huge} is among the classes of
the object in \code{x}.

\code{as.huge()} adds class \code{huge} as the first class of \code{x}.
}
\examples{
as.huge(list(1:100))

}
\seealso{
Other class-related functions provided by ub: 
\code{\link{classes}}
}
\concept{class-related functions provided by ub}
