% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classes}
\alias{classes}
\alias{add_class}
\alias{drop_class}
\title{Add class to the object's classes}
\usage{
add_class(x, class)

drop_class(x, class)
}
\description{
\code{add_class()} adds the class specified in \code{what} as the first
among the classes of the object \code{x}.

\code{drop_class()} removes the specified class from the classes of the
given object.
}
\examples{
add_class(list(1:20), 'huge')

}
\seealso{
Other class-related functions provided by ub: 
\code{\link{class_huge}}
}
\concept{class-related functions provided by ub}
