% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{coalesce_names}
\alias{coalesce_names}
\title{Names functions}
\usage{
coalesce_names(x, y)
}
\description{
\code{coalesce_names()} fills in the missing (zero-length) names in \code{x}
with the corresponding names in \code{y}.

\code{unempty_names()} works similarly except it takes directly
(through \code{names}) the vector names from which to source the
replacements for the missing names in \code{x}.

\code{is_names_empty()} checks for empty names. For an object without
\code{names} (i.e. when \code{names(x)} returns \code{NULL}) it returns a vector of
\code{FALSE} of the same length as \code{x}.

\code{has_any_names()} is an alias for \code{!is.null(names(x))}.

\code{has_all_names()} checks that \code{x} has any names and that all those
names are non-empty strings.
}
\examples{
x = c(a=1, b=2, 3, d=4)
y = c(NA, NA, c=NA, NA)
coalesce_names(x, y)

unempty_names(x, names(y))

is_names_empty(x)
is_names_empty(y)
is_names_empty(1:5)

has_any_names(1:2)
has_any_names(x)

}
\concept{names-related functions provided by ub}
