% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{combine_stouffer}
\alias{combine_stouffer}
\title{P-value combination using Fisher method}
\usage{
combine_stouffer(p, silent = TRUE)
}
\description{
Combines p-values using the Stouffer's combination methods
}
