% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_compare.R
\name{compare_functions}
\alias{compare_functions}
\title{Compare functions}
\usage{
compare_functions(
  fun1,
  fun2,
  by_name = FALSE,
  envir1 = parent.frame(),
  envir2 = parent.frame()
)
}
\description{
Compares two functions by looking at their arguments (via
\code{\link[base:args]{base::args()}}) and their bodies (via \code{\link[base:body]{base::body()}}).
}
\examples{
f1 = function (x) .Internal(which.max(x))
identical(f1, base::which.max)                              # returns FALSE, since they are defined in different environments
compare_functions(f1, base::which.max)                      # returns TRUE
compare_functions('f1', 'base::which.max', by_name=TRUE)    # returns TRUE

}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
