% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{dir_name}
\alias{dir_name}
\title{Extract file name and path}
\usage{
dir_name(files)
}
\description{
\code{dir_name()} is the analogue of \code{base::dirname()} which does not
fail on long paths. \code{base_name()} is the equivalent analogue of
\code{base::basename()}.
}
\examples{
file = 'c:\\\\Windows\\\\System32\\\\cmd.exe'
dir_name(file)     # returns the part part, i.e., 'c:\\Windows\\System32'
base_name(file)    # returns the filename part, i.e., 'cmd.exe'

}
