% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{environment_variables}
\alias{environment_variables}
\alias{get_envir_var}
\alias{set_envir_var}
\title{Get and set environment variables}
\usage{
get_envir_var(name)

set_envir_var(name)
}
\description{
\code{set_envir_var} sets a system environment variable,
\code{get_envir_var} reads a value of a system environment variable.

These are simply wrappers for the built-in R functions so that it
is easier to find them.
}
\seealso{
Other system-related functions provided by ub: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{is_win}()},
\code{\link{stopping_control}},
\code{\link{system_path}},
\code{\link{timest}()},
\code{\link{wait}()}
}
\concept{system-related functions provided by ub}
