% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments}
\alias{environments}
\alias{orig_env}
\alias{parent_envs}
\alias{parent_frames}
\title{Original and parent environments}
\usage{
orig_env(obj, na = "")

parent_envs(x = parent.frame())

parent_frames(n, stop_at_global = FALSE, envir = parent.frame())
}
\description{
\code{orig_env()} gets the "original" environment (which is more precisely
referred to as the \emph{enclosing environment}) of a function (or another
object). As described in the help page of \code{base::parent.env()} (see
\code{?parent.env}), an object's parent environment is the environment where
it was defined. For a function, this is differentiated from the function's
\emph{parent frame}, which is the environment from which the function was invoked.
During object look up, the cascade of parent environments is searched
for an object of the matching name. Parent frames are not part of this
cascade.

When an object defined in a namespace (i.e., inside a package) is printed,
its namespace is shown as the last line. If a function in a package has
been defined as an alias of a function from another package, this original
information is still present when the function is printed. For instance,
\code{ub::head} is just an alias for \code{utils::head}. This can be seen by
calling \code{print(ub::head)}. The function \code{orig_env()} captures that
information and returns it as a character string. For objects with no such
information printed it returns the value in the argument \code{na}.

\code{parent_envs()} returns the cascade of all parent environments
of \code{x}. If no input is specified during a call to \code{parent_envs()},
the default is to take the function's calling environment.

\code{parent_frames()} returns the cascade of parent frames to the frame that
invoked it (unless \code{envir} specifies a different base frame). It returns
a list with the frames and the invoking calls, with the names of the list
frames derived from those calls.
}
\examples{
orig_env(mean)

parent_envs()
parent_envs(mean)

}
