% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_by_positiveness_negativeness}
\alias{extract_by_positiveness_negativeness}
\alias{positive}
\alias{negative}
\alias{nonnegative}
\alias{nonpositive}
\title{Extract elements based on their positiveness/negativeness}
\usage{
positive(x)

negative(x)

nonnegative(x)

nonpositive(x)
}
\description{
\code{positive} extract all positive elements in \code{x}.

\code{negative} extacts all negative elements in \code{x}.

\code{nonpositive} extacts all non-positive elements in \code{x}.

\code{nonnegative} extacts all non-negative elements in \code{x}.
}
\seealso{
Other check-performing functions provided by ub: 
\code{\link{anyNonNA}()},
\code{\link{is_number}()},
\code{\link{is_same_value}()},
\code{\link{numchecks}},
\code{\link{rounding}}
}
\concept{check-performing functions provided by ub}
