% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_read_bytes}
\alias{file_read_bytes}
\title{Read file byte-by-byte}
\usage{
file_read_bytes(
  file,
  what = integer(),
  n,
  size = 1,
  endian = .Platform$endian,
  signed = TRUE
)
}
\description{
Read a file as a binary file. Returns a vector (of type signed integer by default)
with each element corresponding to a single byte in the file (when \code{size=1}).
This is used by \code{file_compare()} for file comparison.
}
