% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{find_column}
\alias{find_column}
\alias{find_column_in_file}
\title{Find a column by name}
\usage{
find_column(
  name,
  df_pattern = ".*",
  fixed = TRUE,
  ignore_case = FALSE,
  envir = parent.frame(1)
)

find_column_in_file(
  name,
  df_pattern = ".*",
  files,
  dir = ".",
  file_pattern = ".*",
  fixed = TRUE,
  ignore_case = FALSE,
  file_ignore_case = TRUE,
  stop_when_found = FALSE,
  nskip = 0
)
}
\description{
\code{find_column()} searches all data frames in the given environment (the parent
frame by default) for a column with the name specified in \code{name}. The matching
of the column name is done exactly (when \code{fixed=TRUE}) or as a regular pattern
(when \code{fixed=FALSE}). The search can be limited to a subset of data frames by
specifying a regular pattern for the name of the data frame via \code{df_pattern}.

\code{find_column_in_file()} performs the same search as \code{find_column()} except in
files specified either explicitely (via \code{files}) or via a pattern (\code{file_pattern}).
Case sensitivity of the file search is controlled by \code{file_ignore_case}.
The search can be interrupted after the first match is found when
\code{stop_when_found=TRUE}. The first \code{nskip} files are skipped.
}
