% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{fix_bibliography}
\alias{fix_bibliography}
\title{Fix LaTeX Bibliography}
\usage{
fix_bibliography(
  infile,
  outfile,
  dot = ".",
  lastname_first = TRUE,
  special_words = "^[Vv]an$|^[Dd]e$|^[Dd]er$|^[Dd]os$"
)
}
\description{
\code{fix_bibliography()} puts author names into the 'Lastname1,
Firstname1 and Lastname2, Firstname2' format in a bibtex bibliography
file \code{infile}. It also abbreviates the first names with/without a dot
and outputs the new bibliography into outfile Assumes that records
are wrapped by '{' and '}' and not '"'. This limitation might be
removed in the future.
}
