% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun_name}
\alias{fun_name}
\alias{this_fun_name}
\alias{parent_fun_name}
\title{Get function name}
\usage{
this_fun_name(e = sys.parent())

parent_fun_name(n = 2L)
}
\description{
\code{this_fun_name()} returns the name of the parent function, i.e. the
function from which it is invoked. If invoked from \code{.GlobalEnv}, it
returns an empty string.

\code{parent_fun_name()} returns the parent function of the parent function.
}
\examples{
example_function = function() print(this_fun_name())
example_function()     # should return 'example_function'

}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
