% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_dump_code}
\alias{function_dump_code}
\title{Dump the source code of a function to a file}
\usage{
function_dump_code(fun, file = stdout())
}
\description{
\code{function_dump_code()} dumps the source code of a function \code{fun}
to a file \code{file} (\code{stdout()} by default).
}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
