% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_find_dependencies}
\alias{function_find_dependencies}
\title{`
Find dependencies of a function}
\usage{
function_find_dependencies(
  fun,
  dep_envir,
  envir = parent.frame(),
  get_status = TRUE,
  announce = TRUE,
  warn = TRUE
)
}
\description{
\code{function_find_dependencies()} searches the source code of a
function (supplied either directly or by name as string via \code{fun})
and identifies invocations of functions that are found in the
specified environment or package (supplied via \code{dep_envir}). When
supplied by name, the environment in \code{dep_envir} is searched for
a function of matching name. The function name can be supplied
together with a package using the double colon notation (e.g.,
'base::mean\verb{), which is then used to set }envir\verb{(unless it has been supplied. When the argument}dep_envir\verb{was not specified, the environment in}envir` is used.
}
\examples{
# Function supplied directly or by name, `envir` is implied (i.e.,
# set to `environment(base::sample)`), `dep_envir` missing so taken 
# the same as `envir`. Thus it looks for dependencies of `sample` 
# in the environment "base".
function_find_dependencies(base::sample)
function_find_dependencies('base::sample') 

# function supplied directly, `envir` is implied (i.e., set to
# `environment(base::sample)`), `dep_envir` is specified explicitly.
function_find_dependencies(base::sample, 'stats')
function_find_dependencies('base::sample', 'stats')
function_find_dependencies('sample', 'stats', 'base')

# An example of a missing function (xxxx)
g = function() print('hi')
ff = function() { x = xxxx(); a = sin(1); g() }
function_find_dependencies(ff)

}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
