% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_modify}
\alias{function_modify}
\alias{function_modify_in_place}
\alias{function_restore}
\alias{function_disable}
\title{Modify and restore functions}
\usage{
function_modify_in_place(
  fun_name,
  calls = "",
  envir = parent.frame(),
  where = c("first", "last", "at"),
  at = NULL,
  restore_call,
  remodify = FALSE,
  quietly = FALSE
)

function_restore(fun_name, envir = parent.frame(), quietly = FALSE)

function_disable(fun_name, envir = parent.frame(), quietly = FALSE)
}
\description{
Functions to \strong{modify in place} and restore other functions located
anywhere including inside locked packages.
}
\details{
\code{function_modify_in_place()} is the workhorse function that can be used to
modify other functions \strong{in place}. It attaches the original source
code of the function that is being modified as the attribute
\code{original_function} of the modified function.

\code{function_restore()} restores the original code of a modified
function (specifically using the attribute \code{original_function}).

\code{function_disable()} can be used to disable any function, i.e., make
the function (supplied by name) do no action. The modification is done
also \strong{in place}.
\code{function_kidnap()} kidnaps the supplied function, where kidnapping
means that the function will ask the user for input before proceeding
with its code. The original code of the kidnapped function is backed
up in the attribute \code{original_function} of the kidnapped function.
The modification is done also \strong{in place}.
}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
