% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hijack}
\alias{hijack}
\title{Argument hijack function}
\usage{
hijack(fun, ...)
}
\description{
A function that allows for changing of default values for
arguments of other functions. Similar to \code{\link[base:formals]{base::formals}}, which it
leverages, but arguably with nicer syntax. Besides revaluing the defaults
of existing arguments, it also allows adding new arguments.
}
\examples{
cat = hijack(cat, sep='')        # equivalent to cat0
mean = hijack(mean, na.rm=TRUE)  # equivalent to mean2

}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by ub}
