% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{is_all_character}
\alias{is_all_character}
\title{Check of class character}
\usage{
is_all_character(x)
}
\description{
\code{is_all_character()} checks whether all elements in an object are of class
\code{character}. For atomic 'x', \code{base::as.character()} is called directly. For
non-atomic \code{x} (e.g. \code{list}), \code{base::as.character()} is called via \code{base::sapply()}.
}
\examples{
is.character(letters)                      # returns TRUE
is.character(list(letters))                # returns FALSE
is_all_character(list(letters))            # returns TRUE
is_all_character(list(letters, 1:10))      # returns FALSE

}
\seealso{
Other list utilities provided by ub: 
\code{\link{is_list_class}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by ub}
