% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is_same_object}
\alias{is_same_object}
\title{Same object check}
\usage{
is_same_object(..., envir = parent.frame())
}
\description{
\code{is_same_object()} checks whether all of the supplied arguments
correspond to the same object in memory. Arguments can be given as
variables or by name (as character).
}
\examples{
a = 1; b = a; c = 2
is_same_object(a, b)     # returns TRUE
is_same_object(a, b, c)  # returns FALSE
is_same_object(a, "a")   # returns TRUE

}
