% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_update}
\alias{list_update}
\alias{modifyList2}
\title{Modify list}
\usage{
list_update(x, val, ..., drop_null_val = TRUE)

modifyList2(x, val, ..., drop_null_val = TRUE)
}
\description{
A version of \code{\link[utils:modifyList]{utils::modifyList()}} which drops zero-length
elements in \code{val} before updating \code{x}. Optionally, it can behave the
exact same way as modifyList (when \code{drop_null_val=FALSE}).
}
\examples{
list_update(list(a=1, b=2), list(a=100, c=200))

}
\seealso{
Other list utilities provided by ub: 
\code{\link{is_all_character}()},
\code{\link{is_list_class}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by ub}
