% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_files.R
\name{load_objects}
\alias{load_objects}
\alias{find_object_in_files}
\alias{list_objects_in_files}
\alias{loadAs}
\title{List and load objects from a file}
\usage{
load_objects(
  file,
  announce = FALSE,
  list_new = FALSE,
  expected_objects = NULL,
  quit_on_miss = FALSE,
  envir = parent.frame()
)

find_object_in_files(
  object_name,
  files = NULL,
  dir = ".",
  pattern = "^.*[.]RData$",
  stop_on_found = TRUE,
  announce = TRUE
)

list_objects_in_files(files = NULL, dir = ".", pattern = "^.*[.]RData$")

loadAs(file, as, what, envir = parent.frame())
}
\description{
\code{list_all_objects()} lists all objects inside given files or
inside files that match pattern in \code{pattern} relative to the path in
\code{dir}.

\code{find_object_in_files()} searches for an object with name in \code{object_name}
inside files in \code{files} relative to the path supplied in \code{dir}. If a
pattern is given (via \code{pattern}) instead of a list of file names,
then all files that match the pattern are searched through looking
for the object in \code{object_name}.

\code{list_objects_in_files()} returns a list of objects in the given files.

\code{load_objects()} is a more verbose version of \code{base::load()}.
\code{load_objects()} can announce which objects have been loaded, the
user can also specify which objects are expected and an error is
thrown if any of these expected objects are not found inside the file.

\code{loadAs()} allows to loads objects from a file and assigns their
contents into variables listed in \code{as} inside the environment in
\code{envir} (the parent frame of the call to \code{loadAs()} by default).
}
