% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{means}
\alias{means}
\alias{mean_na}
\alias{amean}
\alias{hmean}
\alias{gmean}
\alias{gmean2}
\alias{gmeanw}
\alias{lmean}
\title{Various means}
\usage{
mean_na(..., na.rm = TRUE)

amean(x, na.rm = TRUE)

hmean(..., na.rm = TRUE)

gmean(x, na.rm = TRUE)

gmean2(x, na.rm = TRUE, zero.propagate = FALSE)

gmeanw(x, w, na.rm = TRUE, ...)

lmean(x, w, na.rm = TRUE, ...)
}
\description{
Various mean functions. Arithmetic, geometric, weighted geometric,
harmonic and logistic means are implemented. Compares to
\code{\link[base:mean]{base::mean}} these functions have remove \code{NA}s by default and
return zero-length responses on zero-length input or input with no
non-\code{NA} values.

\code{amean()} gives the arithmetic mean of \code{x}.

\code{gmean()} and \code{gmeanw} calculate the geometric mean of \code{x}, while
the latter permits weighing. \code{gmean2()} is a version that processes
inputs with negative values without warnings as well as drops zeros
(unless \code{zero_propagate=TRUE}). \code{gmean2()} inspired by this:
https://stackoverflow.com/a/25555105

\code{hmean()} calculates the harmonic mean (single value) of a single
vector created as a combination of all its input arguments.

\code{lmean()} calculates the (weighted) logistic mean.

\code{mean_na()} is an alias for \code{\link[base:mean]{base::mean}} with \code{na.rm=TRUE}.
}
\examples{
x = c(1,NA); mean(x); mean2(x)

}
