% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{method_dispatch_sequence}
\alias{method_dispatch_sequence}
\title{Method dispatch sequence}
\usage{
method_dispatch_sequence(..., envir = parent.frame())
}
\description{
\code{method_dispatch_sequence} is just an alias for \code{sloop::s3_dispatch} with
an easier to remember name.
}
\examples{
method_dispatch_sequence(print(1))

}
