% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{min_dif}
\alias{min_dif}
\title{Minimal distance within a vector}
\usage{
min_dif(x)
}
\description{
Get the minimal distance between elements of \code{x}.
}
\seealso{
Other numeric functions provided by ub: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by ub}
