% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{mtctestFast}
\alias{mtctestFast}
\title{Fast multiple testing corrections}
\usage{
mtctestFast(
  p,
  alpha = 0.05,
  method = c("bonferroni", "hommel", "holm", "hochberg", "BH", "BY", "fdr", "none",
    "SFG"),
  n1 = -1,
  lam = 2,
  lamSFG = 0.5,
  what = c("fwer", "fdr", "pow"),
  trace = 0,
  dll_lib = NULL,
  unload = TRUE
)
}
\description{
Fast multiple testing corrections
}
