% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{multi_grep}
\alias{multi_grep}
\alias{grepm}
\alias{greplm}
\title{Vectorized regular pattern matching}
\usage{
multi_grep(pattern, x, ..., workhorse = base::grepl)

grepm(pattern, x, ..., exclude = FALSE)

greplm(pattern, x, ..., exclude = FALSE)
}
\description{
Vectorized regular pattern matching, which takes multiple patterns.
}
