% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{nlapply}
\alias{nlapply}
\title{Named list apply}
\usage{
nlapply(...)
}
\description{
\code{nlapply()} is an alias for \verb{base::sapply() which returns a named list by default (by setting }USE.NAMES=TRUE\code{and}simplify=FALSE`).
}
