% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{nlist}
\alias{nlist}
\alias{nlist2}
\title{Named list}
\usage{
nlist(...)

nlist2(...)
}
\description{
\code{nlist()} creates a named list. The code \code{list(a = a, b = b)}
becomes \code{nlist(a,b)} and \code{list(a = a, b = 2)} becomes \code{nlist(a, b =  2)}, etc.

\code{nlist2()} is a shorter original code which relies on existing
code (\code{dots_to_nlist}). It should result in the exact same behavior
as \code{nlist()}.

Credit: The code for \code{nlist()} was lifted from the package \code{loo}
to avoid having to install that package just for this one function.
The code has been slightly modified.
}
\examples{
# uses the variable names to assign names
a = 1:10
nlist(a, b=2)
nlist2(a, b=2)

# uses the call to assign names
nlist(1:10, sin(30))
nlist2(1:10, sin(30))

}
\seealso{
Other list utilities provided by ub: 
\code{\link{is_all_character}()},
\code{\link{is_list_class}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()}
}
\concept{list utilities provided by ub}
