% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_normal.R
\name{normaldist}
\alias{normaldist}
\alias{upnorm}
\alias{uqnorm}
\alias{pmnormrect}
\alias{normpval}
\title{Normal distribution}
\usage{
upnorm(..., lower.tail = FALSE)

uqnorm(..., lower.tail = FALSE)

pmnormrect(
  upper = c(Inf, Inf),
  lower = c(-Inf, -Inf),
  mean = rep(0, length(upper)),
  Sigma = diag(length(upper)),
  varcov
)

normpval(x, two.sided = FALSE, m = 0, s = 1)
}
\description{
\code{upnorm()} returns upper tail probabilities of a normal
distribution.

\code{uqnorm()} returns the upper tail quantiles of a normal
distribution.

\code{pmnormrect()} calculates the probability of an arbitrary box with
respect to an equi-correlated multivariate normal random vector of
dimension given by the size of upper/lower.

\code{normpval()} calculates the p-value relative to the normal
distribution with parameters \code{m} (mean) and \code{s} (standard deviation).
}
\examples{
normpval(1.96)        # approx. 2.5\%
normpval(1.96, TRUE)  # approx. 5.0\%

}
