% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notepad++.R
\name{notepad_key_update}
\alias{notepad_key_update}
\title{Update a single Notepad++ keyword}
\usage{
notepad_key_update(keywords, key_name, code, max_nchar = 20000)
}
\description{
Locates the current keyword (in \code{key_name}) and updates its keywords
with the supplied keywords (\code{keywords}). When the total length of
the key would exceed \code{max_nchar}, new keyword entries are started
and a warning is printed.
}
