% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notepad++.R
\name{notepad_pckg_keys}
\alias{notepad_pckg_keys}
\title{List objects from packages}
\usage{
notepad_pckg_keys(
  pckgs,
  pattern,
  exclude = TRUE,
  what = c("all", "exported"),
  all.names = FALSE
)
}
\description{
Lists the selected objects (all or exported) from the specified packages.
}
