% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notepad.R
\name{notepad_update_highlighting}
\alias{notepad_update_highlighting}
\title{Add all ub objects into Notepad++ highlighting}
\usage{
notepad_update_highlighting(
  fields,
  keywords = NULL,
  npp_path,
  npp_langs,
  backup = TRUE
)
}
\description{
Modifies the notepad language keyword groups in order for the
objects in 'ub' to be highlighted by Notepad++.
}
