% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{null}
\alias{null}
\title{Null function (function that does nothing)}
\usage{
null(...)
}
\description{
\code{null()} is a function that takes arbitrary arguments and does nothing.
}
\seealso{
Other coding-related functions provided by ub: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{function_separate_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()}
}
\concept{coding-related functions provided by ub}
