% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_NULL}
\alias{operator_NULL}
\alias{\%|||\%}
\alias{\%||||\%}
\alias{\%|||||\%}
\alias{\%NA\%}
\alias{\%ERRCLS\%}
\alias{\%ERR\%}
\title{Default value for NULL and zero-length objects}
\usage{
x \%|||\% y

x \%||||\% y

x \%|||||\% y

x \%NA\% y

x \%ERRCLS\% y

x \%ERR\% y
}
\description{
\verb{\\\%|||\\\%} is intented to be equivalent "\code{\%||\%}" operator in
the package \code{rlang}, but named differently as not to clash with it.

\verb{\\\%||||\\\%} extends this behavior to all zero-length objects.

\verb{\\\%|||||\\\%} extends this behavior further to include empty strings.

\verb{\\\%NA\\\%} is an operator version of `ifelse(is.na(x), y, x), which
allows input of various types (vectors, lists, closures, etc.).

\verb{\\\%ERR\\\%} works analogously except that it checks for class
\code{try-error}. If its first argument is of class \code{try-error} it returns
its second argument (\code{y}). Otherwise it returns the first argument (\code{x}).

\verb{\\\%NEXISTS\\\%} works analogously except that it avoids an error due to
the non-existence of the first argument by returning the second argument
instead.
}
\examples{
1 \%|||\% 2                                   # returns 2
NULL \%|||\% 2                                # returns 2
NULL \%||||\% 2                               # returns 2
numeric(0) \%||||\% 2                         # returns 2
round('a') \%ERR\% 0                          # returns 0
try(round('a'), silent=TRUE) \%ERRCLS\% 0     # returns 0
.name.of.a.nonexisting.object. \%NEXISTS\% 0  # returns 0

}
\seealso{
Other operators provided by ub: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{not_in}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}},
\code{\link{renaming_operators}}
}
\concept{operators provided by ub}
