% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{palette_ub}
\alias{palette_ub}
\title{Color palette}
\usage{
palette_ub(n)
}
\description{
Returns a specified number of colors a from a color palette. If
the requested number of colors is larger than the size of the palette
the colors are recycled.
}
\examples{
palette_ub()
palette_ub(3)

}
\seealso{
Other plotting-related functions provided by ub: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by ub}
