% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{path_8dot3}
\alias{path_8dot3}
\alias{path_to_8dot3}
\alias{path_from_8dot3}
\title{Convert to and from 8dot3 paths}
\usage{
path_to_8dot3(path, mustWork = TRUE)

path_from_8dot3(path, winslash = "/", mustWork = NA)
}
\description{
\code{path_to_8dot3()} converts and existing path to the 8dot3 format.
It basically does the same thing as the built-in \code{utils::shortPathName()}
(slower but directly via shell).

\code{path_from_8dot3()} does the opossite conversion. Keep in mind though
that \code{path_from_8dot3()} is not an exact inverse operation and the
results can differ from the starting path in a number of ways
(e.g., type of directory separator, trailing slash, capitalization).
}
\examples{
path = 'c:/program files/'
path_short = print(path_to_8dot3(path))
path_orig = path_from_8dot3(path_short)

}
