% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{read_zip_single}
\alias{read_zip_single}
\title{Read a table from a single zip archive}
\usage{
read_zip_single(
  file,
  zipfile,
  zip_type,
  list_connections,
  fun_read,
  trace = 0,
  ...
)
}
\description{
Read a table from a single zip archive
}
