% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode2}
\alias{recode2}
\title{Recode values function}
\usage{
recode2(.x, conv, .default = NULL, .missing = NULL, .fun = dplyr::recode)
}
\description{
A slightly modified version of \code{dplyr::recode}, which takes
the recoding specification as a named vector.
}
\examples{
recode2(letters[1:10], c('a'='Aaron','b'='Brenda'))
recode2(letters, a, 'b'='BBB', .default='<none>')

}
