% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalue.R
\name{replace_value}
\alias{replace_value}
\alias{unzero}
\alias{unspace}
\alias{de_na}
\alias{bound_between}
\title{Replace values}
\usage{
unzero(x, value = 1, zero = 0)

unspace(x, s = "-", space = "\\\\s+", workhorse = gsub)

de_na(x, y, use_all_y = FALSE)

bound_between(x, lower, upper, na.rm = FALSE)
}
\description{
\code{unzero()} replace all zeros in \code{x} with the content of \code{value}.

\code{unspace()} replaces  all space characters (or characters matching
the pattern supplied in \code{space}) with dashes (or the value supplied
in \code{s}).

\code{de_na()} replaces all \code{NA} elements in its first argument (\code{x})
with the corresponding values in its second argument (\code{y}). Same
effect as \code{dplyr::coalesce}, except when \code{use_all_y} is \code{TRUE},
when the \code{NA} elements in \code{x}, say there are \\emph{n} of them, are
replaced with the first \emph{n} elements of \code{y} (with possible
recycling of \code{y}).

\code{bound_between()} between bounds the values in \code{x} by \code{lower} from
below and \code{upper} from above. In other words, all elements in \code{x}
that exceed \code{lower} or \code{upper} are replaced with the bounds.
}
\examples{
unzero(c(0,0,10,10,0))
unzero(c(0,0,10,10,0), NA)

unspace('this is a beautiful world.')
unspace('this is a beautiful world.', '_')
de_na(c(1,2,NA,4,5,NA), 0)
de_na(c(1,2,NA,4,5,NA), 1:6)
de_na(c(1,2,NA,4,5,NA), Inf)

bound_between(-10:10, 0, 5)
bound_between(-4:5, rep(c(-1,-2),5), rep(

}
\seealso{
Other numeric functions provided by ub: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by ub}
