% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rnormpval}
\alias{rnormpval}
\title{Draw random sample of p-values under the normal distribution}
\usage{
rnormpval(
  M,
  mu,
  Sigma = def_Sigma(rho, N = length(mu)),
  rho = 0,
  two.sided = FALSE,
  trace = 1,
  what = c("p", "X", "pX")
)
}
\description{
\code{rnormpval()} draws a random sample of size \code{M} of p-values
relative to the normal distribution with parameters \code{m} (mean) and
\code{s} (standard deviation).
}
