% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_update.R
\name{rows_modify_recurse}
\alias{rows_modify_recurse}
\alias{rows_patch_recurse}
\alias{rows_upsert_recurse}
\alias{rows_insert_recurse}
\alias{rows_delete_recurse}
\alias{rows_update_recurse}
\title{Recursive update}
\usage{
rows_modify_recurse(d, ..., fun)

rows_patch_recurse(d, ...)

rows_upsert_recurse(d, ...)

rows_insert_recurse(d, ...)

rows_delete_recurse(d, ...)

rows_update_recurse(d, ...)
}
\description{
\code{rows_modify_recurse()} iteratively alters the elements of a list of data
frames by applying \code{fun()} starting with the first element in the list and
going through the rest. The intended values for \code{fun} are \code{dplyr::rows_upsert},
\code{dplyr::rows_update}, \code{dplyr::rows_insert}, \code{dplyr::rows_patch} and
\code{dplyr::rows_delete} but can be any function that takes two dataframes.
}
\details{
\code{rows_upsert_recurse()}, \code{rows_update_recurse()}, \verb{rows_insert_recurse(), }rows_patch_recurse()\code{and}rows_delete_recurse()\verb{are wrappers around}rows_modify_recurse()\verb{with the dplyr functions in the place of}fun`.
}
