% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{samplepdf}
\alias{samplepdf}
\title{Sample from a distribution}
\usage{
samplepdf(n, pdf, ..., spdf.lower = -Inf, spdf.upper = Inf)
}
\description{
\code{samplepdf()} samples \code{n} variables from an arbitrary density
function specified in \code{pdf}.
}
