% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveRDSlazy}
\alias{saveRDSlazy}
\title{Save to file}
\usage{
saveRDSlazy(
  object,
  file = "",
  backup = TRUE,
  check_for_sameness = TRUE,
  announce = TRUE,
  ...
)
}
\description{
A lazy version \code{base::saveRDS} that writes to the specified file but only
when the file does not yet exist or when the contents of the existing file
would be changed by writing \code{object} into it.
}
\seealso{
Other file-reading functions provided by ub: 
\code{\link{read_char}()}
}
\concept{file-reading functions provided by ub}
