% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{separate_path}
\alias{separate_path}
\title{Split file names}
\usage{
separate_path(files, path0 = "./")
}
\description{
Separates the path portions from file names. Returns the paths and
the proper file names in a list.
}
\examples{
separate_path("c:/Windows/system.dat")
separate_path("system.dat")

}
\seealso{
Other file system function provided by ub: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_backup}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{map_drives}},
\code{\link{random_filename}()}
}
\concept{file system function provided by ub}
