% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_around}
\alias{seq_around}
\title{Produce a sequence of numbers}
\usage{
seq_around(a, b, m1, m2, len, add_a = FALSE, add_b = FALSE)
}
\description{
Produces an equidistant sequence between \code{a} and \code{b} of length approximately
\code{len} which contains both \code{m1} and \code{m2}. The sequence might not contain the
border points \code{a} and \code{b}. Enforcing their presence in the sequence can be
done via  \code{add_a=TRUE} and/or \code{add_b=TRUE}, althought then the equi-distance
of all points might no longer be true.
}
\examples{
seq_around(0, 10, 7, len=20)

}
\seealso{
Other sequence-related functions provided by ub: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by ub}
