% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{short_notation_exp}
\alias{short_notation_exp}
\title{Convert a number to an abbreviated notation using}
\usage{
short_notation_exp(x, base = 10, ndig = 2, as_expr = FALSE)
}
\description{
Convert a number to an abbreviated notation using
}
