% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_matrix}
\alias{simulate_matrix}
\alias{rposdefmat}
\alias{rvariance}
\title{Draw a positive-definite matrix}
\usage{
rposdefmat(n, ev = runif(n, 1, 10), all.positive = FALSE)

rvariance(n, sigma2 = rep(1, n), ev = runif(n, 1, 200))
}
\description{
\code{rposdefmat()} generates a \"random\" (more like arbitrary)
positive-definite matrix with user-specified positive eigenvalues. If
eigenvalues are not specified, they are generated from a uniform
distribution on (1,10)

\code{rvariance()} draws a \"random\" variance matrix with \code{sigma2} on
the diagonal.
}
