% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_pval}
\alias{simulate_pval}
\title{Simulate p-values}
\usage{
simulate_pval(
  M,
  Sigma,
  N,
  n = 1,
  n0 = N,
  mu0 = 0,
  n1 = 0,
  mu1 = 0,
  alpha = 0.05,
  type = "bySigma",
  model = c("norm", "unif"),
  two.sided = FALSE,
  evaluate_signif = FALSE,
  lambda,
  trace = 1
)
}
\description{
\code{simulate_pval()} simulates p-values (from the normal model only
so far).
}
