% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ub.R
\name{source_ub}
\alias{source_ub}
\title{Source ub code files}
\usage{
source_ub(..., path, normalize = FALSE, envir = .GlobalEnv, report_new = TRUE)
}
\description{
Sources the specified files relative to the path in \code{path}. If
\code{files} is not specified, it sources all files in the path.

By default, this sources files from the ub location on my
system (the author).
}
\seealso{
Other ub-internal functions provided by ub: 
\code{\link{ub_internals}}
}
\concept{ub-internal functions provided by ub}
