% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{stop2}
\alias{stop2}
\alias{stop_quietly}
\title{Stop execution without the error message produced by \code{base::stop}.}
\usage{
stop2(msg = "Execution stopped.", show_sequence = TRUE)

stop_quietly(msg = "Execution stopped.", show_sequence = TRUE)
}
\description{
Stops the execution of code without an error message such as the
one produced by \code{base::stop}. By default, prints a message about
stopping quietly with the calling sequence. To stop completely
silently, use \code{stop_quietly(NULL)}.
}
