% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_break_at_space}
\alias{str_break_at_space}
\title{Break string into lines (at white spaces only)}
\usage{
str_break_at_space(
  x,
  max_width = Inf,
  drop_space_after_newline = FALSE,
  eol = "\\n"
)
}
\description{
Break strings into lines so that the maximum length of a line does
not exceed \code{max_width} except when there is a single word longer
than \code{max_width}). Lines are places only at white spaces.
}
