% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_script.R
\name{strip_start}
\alias{strip_start}
\title{\code{strip_start()} removes the strings that indicate help lines (i.e., "#'").}
\usage{
strip_start(x)
}
\description{
\code{strip_start()} removes the strings that indicate help lines (i.e., "#'").
}
