% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{t_day}
\alias{t_day}
\title{Get today's date}
\usage{
t_day(format = "\%Y-\%m-\%d", fun = base::as.character)
}
\arguments{
\item{format}{Format of the output.}

\item{set_return_class}{A function that determines the class of
the returned date. Uses \code{base::as.character} by default but for
full dates, while \code{base::as.Date} can be also useful.}
}
\description{
Returns today's date (in the current time zone) in the specified
format.
}
\examples{
t_day()     # current day
t_day('\%Y') # current year
t_day('\%m') # current month

}
\seealso{
Other calendar functions provided by ub: 
\code{\link{closest_day}()},
\code{\link{y_day}()}
}
\concept{calendar functions provided by ub}
