% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{test_compare_models}
\alias{test_compare_models}
\title{Test models for difference}
\usage{
test_compare_models(m1, m2, df = 1)
}
\description{
Calculates the p-value of a chisquare test of difference between
the log-likelihoods of two models (\code{m1} and \code{m2}) at \code{df} degrees
of freedom.
}
