% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timners}
\alias{timners}
\alias{start_timer}
\alias{stop_timer}
\alias{read_timer}
\alias{stop_all_timers}
\title{Timing of runtime duration}
\usage{
start_timer(
  timer_name = NULL,
  announce = TRUE,
  envir = ub_environment(),
  lead = NULL,
  what = "timer",
  out = NULL,
  format = identity
)

stop_timer(
  timer_name = NULL,
  announce = TRUE,
  envir = ub_environment(),
  lead = NULL,
  what = "timer",
  out = NULL,
  format = identity
)

read_timer(
  timer_name = NULL,
  announce = TRUE,
  start = FALSE,
  envir = ub_environment()
)

stop_all_timers(envir = ub_environment())
}
\description{
\code{start_timer()}, \code{stop_timer()}, \code{read_timers()}, \code{stop_all_timers()} can
be used to measure the length of duration of operations. Multiple nested
timers are supported. \code{start_timer()} starts a timer (the name of the timer
can be specified), while \code{stop_timer()} stops either a specific timer
(when its name is supplied via \code{timer_name}) or the most recently started
timer. \code{stop_all_timers()} stops all existing timers. When a timer is
stopped, the duration information is printed. A current duration by a
specific timer can be obtained via \code{read_timer()}.
}
