% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbase.R
\name{to_base_k}
\alias{to_base_k}
\alias{int2k}
\title{Convert base}
\usage{
to_base_k(
  x,
  k,
  recode = TRUE,
  min_ndig = 1,
  codes = c(0:9, letters, LETTERS),
  sep = NULL,
  collapse = NULL,
  simplify = TRUE
)

int2k(x, k, ..., simplify = TRUE)
}
\description{
\code{to_base_k()} converts the values in \code{x} (assumed to be \code{numeric},
and thus in base 10) from the decimal base to base \code{k}.

\code{int2k()} is a version of \code{to_base_k()}, which allows vector input
in \code{k}, and returns the conversion of all values in \code{x} into all
bases specified in \code{k}.
}
\seealso{
Other numeric functions provided by ub: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by ub}
