% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ub.R
\name{ub_internals}
\alias{ub_internals}
\alias{ub_namespace}
\alias{ub_environment}
\alias{get_in_ub_env}
\alias{assign_in_ub_env}
\alias{update_in_ub_env}
\alias{init_in_ub_env}
\alias{get_ub_var_inits}
\title{Utilbox internals}
\usage{
ub_namespace()

ub_environment(envir = .GlobalEnv)

get_in_ub_env(nam, envir = ub_namespace())

assign_in_ub_env(nam, val, envir = ub_namespace())

update_in_ub_env(nam, val, operation = `+`, envir = ub_namespace())

init_in_ub_env(nams, envir = ub_namespace())

get_ub_var_inits(nam)
}
\description{
\code{get_in_ub_env} retrieves a value of a variabled named in \code{x} from
the ub's package namespace or the hidden environment created in
\code{.GlobalEnv} by a call to \code{ub_environment()}.

\code{assign_in_ub_env} is an analogue which assigns a supplied value
(\code{val}) to a variable (\code{x}).

\code{update_in_ub_env} updates the existing variable by applying a
function (\code{operation}) with the variable \code{x} and \code{val} supplied to it
as the first two arguments.

\code{init_in_ub_env} initializes a given variable to its default value
(defined in and obtained via \code{get_ub_var_inits}). If not variable name
is supplied, all variables defined in \code{get_ub_var_inits} are initialized.

\code{ub_namespace()} returns the environment which is the
namespace of the ub package. If it failed for some reason, it
returns creates and/or returns the hidden environment located in
\code{.GlobalEnv}.

\code{ub_environment()} creates the working environment called
\code{.ub}, which is used by some of the functions of the \code{ub}
package (e.g. \code{cat0}) to store private variables.
}
\seealso{
Other ub-internal functions provided by ub: 
\code{\link{source_ub}()}
}
\concept{ub-internal functions provided by ub}
