% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait}
\alias{wait}
\alias{sleep}
\alias{ask_to_confirm}
\alias{ask_to_confirm_with_timeout}
\alias{ask}
\title{Pause execution / wait}
\usage{
wait(..., timeout = Inf)

sleep(time, announce = TRUE)

ask_to_confirm(msg = "Press <ENTER> to continue or <ESC> to quit ...", ...)

ask_to_confirm_with_timeout(msg = "", timeout = 10, no = NULL)

ask(
  question,
  answers = c("Y", "N"),
  affirmative = "Y",
  nmax = 1,
  case_sensitive = FALSE
)
}
\description{
\code{wait()} pauses the execution until either the <ENTER> key (for
continue) or <ESC> (for quit) are pressed.

\code{sleep()} pauses execution for \code{time} seconds with (optional)
announcement.

\code{ask()} pauses the execution until and waits for the user to
respond with one of the questions. It then evaluates whether the
response was in the affirmative (as specified via \code{affirmative}).
Only the first nmax characters of the response are compared with

\code{ask_to_confirm()} and \code{ask_to_confirm_with_timeout()} ask the user for
confirmation with the given message. For the latter, a lack of input within
\code{timeout} seconds is considered non-confirmation, and so are the answers
listed in \code{no}.
}
\examples{
sleep(3)

wait("It's nice weather outside.")

wait("It's nice weather outside.")

}
\seealso{
Other system-related functions provided by ub: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{stopping_control}},
\code{\link{system_path}},
\code{\link{timest}()}
}
\concept{system-related functions provided by ub}
