% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_wchisq.R
\name{weighted_chisquare}
\alias{weighted_chisquare}
\alias{qchisqw}
\alias{pchisqw}
\title{Weighted chi-square distribution}
\usage{
qchisqw(p, weights, lower.tail = TRUE, epsw = 1e-05, epsv = 1e-12, df0 = 2)

pchisqw(x, weights, lower.tail = TRUE, epsw = 1e-05, df0 = 2)
}
\description{
\code{qchisqw()} calculates the quantile function of the weighted
chisquare distribution, which is the distribution that arises when
independent chisquare-distributed variables with \code{df0} degrees of
freedom are summed on the logarithmic scale with weights in \code{w}.

\code{pchisqw()} calculates the probability function of that same
distribution. It gives the left tail (lower.tail==TRUE) or the right
tail (lower.tail=FALSE) probability. The weights are specified via
\code{weights}. When some (but not all!) weights are equal, it gives an
error.
}
