% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{which_package}
\alias{which_package}
\title{Identify object's package}
\usage{
which_package(object_name)
}
\description{
\code{which_package} tries to identity the package in which the
supplied object is located or from which it came from.
}
\seealso{
Other package-related functions provided by ub: 
\code{\link{apply_pckg}()},
\code{\link{auto_install_packages}},
\code{\link{lib_dependencies}()},
\code{\link{list_installed_packages}()},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by ub}
