% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{zip_files}
\alias{zip_files}
\title{Compress multiple files into single archive}
\usage{
zip_files(files, path = ".", zipfile, extras = "-m")
}
\arguments{
\item{files}{List of files to compress. Exact file names, not
regular patterns.}

\item{path}{Path to the files.}

\item{zipfile}{Name of the zip archive to create.}

\item{extras}{Modifiers sent to the zip archiver. Defaults to
'-m', which moves the files into the archive.}
}
\value{
List of file names of the created archive and the
corresponding status codes returned by the archiver.
}
\description{
\code{zip_files} compresses files into a single archive supplied via
\code{zipfile}.
}
\seealso{
Other compression-related utilities provided by ub: 
\code{\link{list_zip}()},
\code{\link{read_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_all_in_path}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_file}()}
}
\concept{compression-related utilities provided by ub}
